requirements.txt
asgiref==3.5.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Django==4.0.2
django-debug-toolbar==3.2.4
django-elasticsearch-dsl==7.2.2
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
elasticsearch==7.17.0
elasticsearch-dsl==7.4.0
factory-boy==3.2.1
Faker==13.0.0
h11==0.13.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
mypy-extensions==0.4.3
outcome==1.1.0
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.0.1
pytest-base-url==1.4.2
pytest-django==4.5.2
pytest-factoryboy==2.1.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-selenium==2.0.1
pytest-variables==1.9.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tenacity==6.3.1
tomli==2.0.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
wsproto==1.0.0
